# backend/src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: HopeCash API
  description: |
    API REST da plataforma HopeCash (NewHope) — gestão financeira pessoal e familiar,
    offline-first. Todas as entidades sincronizáveis carregam `id` (UUID), `version`,
    `updated_at` e `deleted_at` (soft delete). Erros seguem o envelope
    `{ "error": { "code", "message", "details?" } }`.
  version: 1.0.0
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /transactions:
    get:
      tags: [Transactions]
      summary: Lista lançamentos
      parameters:
        - { $ref: '#/components/parameters/updatedSince' }
        - { name: type, in: query, schema: { type: string, enum: [income, expense, transfer] } }
        - { name: status, in: query, schema: { type: string, enum: [planned, paid, overdue, canceled] } }
        - { name: account_id, in: query, schema: { type: string } }
        - { name: category_id, in: query, schema: { type: string } }
      responses: { '200': { description: OK } }
    post:
      tags: [Transactions]
      summary: Cria lançamento (receita/despesa)
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/Transaction' }
      responses: { '201': { description: Criado } }
  /transactions/{id}:
    get: { tags: [Transactions], summary: Detalhe, parameters: [{ $ref: '#/components/parameters/id' }], responses: { '200': { description: OK } } }
    put: { tags: [Transactions], summary: Atualiza, parameters: [{ $ref: '#/components/parameters/id' }], responses: { '200': { description: OK }, '409': { description: Conflito de versão } } }
    delete: { tags: [Transactions], summary: Soft delete, parameters: [{ $ref: '#/components/parameters/id' }], responses: { '200': { description: OK } } }
  /transactions/installments:
    post:
      tags: [Transactions]
      summary: Compra parcelada (cria N parcelas mensais)
      requestBody:
        content:
          application/json:
            schema:
              allOf:
                - $ref: '#/components/schemas/Transaction'
                - type: object
                  required: [total_amount, installments, first_due_date]
                  properties:
                    total_amount: { type: number }
                    installments: { type: integer, minimum: 2 }
                    first_due_date: { type: string, format: date }
      responses: { '201': { description: Parcelas criadas } }
...
  /sync/push:
    post:
      tags: [Sync]
      summary: Envia lote de operações offline (idempotente)
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [device_id, operations]
              properties:
                device_id: { type: string }
                operations:
                  type: array
                  items:
                    type: object
                    required: [operation_id, entity, entity_id, op]
                    properties:
                      operation_id: { type: string }
                      entity: { type: string }
                      entity_id: { type: string, format: uuid }
                      op: { type: string, enum: [create, update, delete] }
                      payload: { type: object }
                      base_version: { type: integer }
                      client_updated_at: { type: string }
      responses:
        '200': { description: 'Resultado por operação: applied | conflict_resolved | duplicate | rejected' }
  /sync/pull:
    get:
      tags: [Sync]
      summary: Alterações incrementais desde o cursor
      parameters:
        - { name: since, in: query, schema: { type: string }, description: 'Cursor do pull anterior; 0 para bootstrap' }
        - { name: entities, in: query, schema: { type: string }, description: 'Lista separada por vírgula (opcional)' }
      responses:
        '200': { description: Registros alterados por entidade + novo cursor }

  # ── Notificações push (Firebase Cloud Messaging) ────────────────────────
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
    id:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    updatedSince:
      name: updated_since
      in: query
      schema: { type: string }
      description: Retorna apenas registros alterados após esse carimbo
  schemas:
...
    Transaction:
      type: object
      required: [type, description, competence_date]
      properties:
        id: { type: string, format: uuid, description: 'UUID gerado no cliente (offline-first)' }
        type: { type: string, enum: [income, expense, transfer] }
        description: { type: string }
        amount_planned: { type: number }
        amount: { type: number }
        competence_date: { type: string, format: date }
        due_date: { type: string, format: date }
        payment_date: { type: string, format: date }
        status: { type: string, enum: [planned, paid, overdue, canceled] }
        account_id: { type: string, format: uuid }
        card_id: { type: string, format: uuid }
        category_id: { type: string, format: uuid }
        subcategory_id: { type: string, format: uuid }
        tags: { type: array, items: { type: string } }
        notes: { type: string }